Added phrase provider
