+++ HW1/Evgeniy_Nikulin/Doors_and_levels_game/Doors_and_levels_game/Resources/ruLanguage.json
@@ -1,0 +1,8 @@
+{
+  "Welcome": "The game of doors\n\nДобро пожаловать в мой замок, тебе необходимо выбрать верную дверь.\n",
+  "YourChoose": "Твой выбор: ",
+  "MoveForward": "Ты двигаешься вперед через выбранную дверь",
+  "MoveBack": "Ты возвращаешься к предыдущим дверям",
+  "YouWin": "\nТы выйграл!",
+  "WrongValue": "Посмотри, тут есть дверь как эта?\nВот я тоже не вижу, попробуй снова."
+}
